fastboot: add -R, --reboot switch
Add a fastboot switch to reboot after finishing an operation. This
allows a user to do `fastboot flash system -R' and reboot the device
after the system partition flashes.

The existing reboot *command* cannot be conveniently chained with a
flash command the same way -w can be combined with flash[all].

Change-Id: I4e9f1e7b6336369bc35150f92ff9a56b3db2ef42
Signed-off-by: Adi <adarsh.mmx@micromaxinfo.com>